bug java final fields
git-svn-id: https://www.lri.fr/svn/demons/why2/trunk@4006 ecafba89-7d34-45a1-8a0b-92829e960b72

+++ lib/javacard_api/com/sun/javacard/impl/Constants.java
@@ -4,6 +4,6 @@ package com.sun.javacard.impl;
 
 public interface Constants {
 
-  public static final int APDU_BUFFER_LENGTH;
+  public static final short APDU_BUFFER_LENGTH;
 
 }
